added unit test and search in service for water ads

+++ bea.domain/Ads/WaterSport/MotorBoatAd.cs
@@ -9,7 +9,7 @@ namespace Bea.Domain.Ads.WaterSport
     public class MotorBoatAd : BaseAd
     {
         public virtual MotorBoatType MotorBoatType { get; set; }
-        public virtual Decimal Length { get; set; }
+        public virtual Double Length { get; set; }
         public virtual int Year { get; set; }
         public virtual MotorBoatEngineType MotorType { get; set; }
         public virtual int Hp { get; set; }
